use uint for beamSizes

+++ src/amun/common/base_best_hyps.h
@@ -21,7 +21,7 @@ public:
       const Words& filterIndices,
       bool returnAlignment,
       std::vector<Beam>& beams,
-      std::vector<size_t>& beamSizes
+      std::vector<uint>& beamSizes
       ) = 0;
 
 };
