+++ src/amun/gpu/dl4mt/decoder.h
@@ -140,7 +140,7 @@ class Decoder {
                                      const mblas::Matrix& HiddenState,
                                      const mblas::Matrix& SourceContext,
                                      const DeviceVector<int>& mapping,
-                                     const std::vector<size_t>& beamSizes)
+                                     const std::vector<uint>& beamSizes)
         {
           // mapping = 1/0 whether each position, in each sentence in the batch is actually a valid word
           // batchMapping = which sentence is each element in the batch. eg 0 0 1 2 2 2 = first 2 belongs to sent0, 3rd is sent1, 4th and 5th is sent2
@@ -338,7 +338,7 @@ class Decoder {
                   const mblas::Matrix& Embeddings,
                   const mblas::Matrix& SourceContext,
                   const DeviceVector<int>& mapping,
-                  const std::vector<size_t>& beamSizes)
+                  const std::vector<uint>& beamSizes)
     {
       //std::cerr << "State=" << State.Debug(1) << std::endl;
       //std::cerr << "Embeddings=" << Embeddings.Debug(1) << std::endl;
@@ -407,7 +407,7 @@ class Decoder {
                                   const mblas::Matrix& HiddenState,
                                   const mblas::Matrix& SourceContext,
                                   const DeviceVector<int>& mapping,
-                                  const std::vector<size_t>& beamSizes) {
+                                  const std::vector<uint>& beamSizes) {
       alignment_.GetAlignedSourceContext(AlignedSourceContext, HiddenState, SourceContext,
                                          mapping, beamSizes);
     }
